+++ Code/SPath/src/kaloffl/spath/tracing/Ray.scala
@@ -1,7 +1,11 @@
 package kaloffl.spath.tracing
 
-import kaloffl.spath.math.Vec3f
+import kaloffl.spath.math.Vec3d
 
-class Ray(val start: Vec3f, val normal: Vec3f) {
-//  val normal = direction.normalize
+/**
+ * A ray consists of a staring point and a direction.
+ *
+ * @author Lars Donner
+ */
+class Ray(val start: Vec3d, val normal: Vec3d) {
 }
